integrate product_selector in enter_orders menu
not the product_selector-widget is accessible through the
enter_orders_menu from mainWindow's-menuBar.
It is also noticeable, that i made changes on the qstditemmodel
undoMechanism and (hopefully) finally fixed the DragDrop MacroCmd.
Moreover hplan_model was extended, to keep track of all orders
(scheduled and unscheduled/stashed separately).
This is mainly to being able to provide OrderLists to enter_won_popup,
(which needs to prevent duplicate WON's in the model)
and wonListView ( whose searchWON-lineEdit's completer needs it).
The enterWON-Dialog now asks for won and quantity.